Not allowed to copy blocks which contain free
variables

+++ core/blockly.js
@@ -262,6 +262,11 @@ Blockly.copy_ = function(toCopy) {
   if (toCopy.isComment) {
     var xml = toCopy.toXmlWithXY();
   } else {
+    if (!toCopy.resolveReference(null)) {
+      // TODO(harukam): Tell users why the block can not be copied.
+      Blockly.alert('This blocks can be copied.');
+      return;
+    }
     var xml = Blockly.Xml.blockToDom(toCopy);
     // Copy only the selected block and internal blocks.
     Blockly.Xml.deleteNext(xml);
